+++ source/common/util/table_2D.inl
@@ -4,8 +4,8 @@ namespace nbl { namespace util {
 
 template<typename T, bool gpu_flag>
 CPU table_2D<T, gpu_flag> table_2D<T, gpu_flag>::create(
-	real x_min, real x_max, size_t width,
-	real y_min, real y_max, size_t height,
+	real x_min, real x_max, int width,
+	real y_min, real y_max, int height,
 	T* data)
 {
 	table_2D<T, gpu_flag> table;
@@ -59,13 +59,13 @@ CPU void table_2D<T, gpu_flag>::set(table_2D<T, other_gpu_flag> const & source)
 }
 
 template<typename T, bool gpu_flag>
-PHYSICS T & table_2D<T, gpu_flag>::operator()(size_t i, size_t j)
+PHYSICS T & table_2D<T, gpu_flag>::operator()(int i, int j)
 {
 	return *(reinterpret_cast<T*>(reinterpret_cast<uint8_t*>(_data) + i * _pitch) + j);
 }
 
 template<typename T, bool gpu_flag>
-PHYSICS T const & table_2D<T, gpu_flag>::operator()(size_t i, size_t j) const
+PHYSICS T const & table_2D<T, gpu_flag>::operator()(int i, int j) const
 {
 	return *(reinterpret_cast<T*>(reinterpret_cast<uint8_t*>(_data) + i * _pitch) + j);
 }
@@ -114,13 +114,13 @@ PHYSICS T table_2D<T, gpu_flag>::get_nearest(real x, real y) const
 }
 
 template<typename T, bool gpu_flag>
-PHYSICS size_t table_2D<T, gpu_flag>::width() const
+PHYSICS int table_2D<T, gpu_flag>::width() const
 {
 	return _width;
 }
 
 template<typename T, bool gpu_flag>
-PHYSICS size_t table_2D<T, gpu_flag>::height() const
+PHYSICS int table_2D<T, gpu_flag>::height() const
 {
 	return _height;
 }
@@ -148,7 +148,7 @@ namespace detail
 	template<typename T>
 	struct table_2D_factory<T, false>
 	{
-		inline static CPU void allocate(table_2D<T, false> & table, size_t width, size_t height)
+		inline static CPU void allocate(table_2D<T, false> & table, int width, int height)
 		{
 			table._width = width;
 			table._height = height;
@@ -181,7 +181,7 @@ namespace detail
 		{
 			// Make indirect array
 			T** host_pp = new T*[table._width];
-			for (size_t x = 0; x < table._width; ++x)
+			for (int x = 0; x < table._width; ++x)
 				host_pp[x] = reinterpret_cast<T*>(reinterpret_cast<uint8_t*>(table._data) + x * table._pitch);
 
 			callback(host_pp);
@@ -203,7 +203,7 @@ namespace detail
 	template<typename T>
 	struct table_2D_factory<T, true>
 	{
-		inline static CPU void allocate(table_2D<T, true> & table, size_t width, size_t height)
+		inline static CPU void allocate(table_2D<T, true> & table, int width, int height)
 		{
 			table._width = width;
 			table._height = height;
